Added new test cases and clear _data after each query
When calling the Singleton Vulnerability_query, the _data variable was not cleared after each query. The result was that the data from one query carries over to the next query. This change could fix that. I also added more test cases to the test class for diffrerent package version IDs